src/utils/requirements.txt
anthropic==0.23.1
pandas==2.1.0
pdf2image==1.17.0
selenium==4.18.0
requests==2.31.0
pathlib==1.0.1
poppler-utils==0.1.0
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
black==23.3.0
ruff==0.0.262
mypy==1.3.0
pre-commit==3.3.2
pytest==7.3.1
